refactor: A proper Rust library structure.
