+++ src/set/clone.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2023-2025 Yegor Bugayenko
 // SPDX-License-Identifier: MIT
 
-use crate::Set;
+use super::Set;
 
 impl<T: Clone, const N: usize> Clone for Set<T, N> {
     fn clone(&self) -> Self {
