+++ src/set/display.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2023-2025 Yegor Bugayenko
 // SPDX-License-Identifier: MIT
 
-use crate::Set;
+use super::Set;
 use core::fmt::{self, Display, Formatter, Write};
 
 impl<T: Display, const N: usize> Display for Set<T, N> {
