+++ src/set/drain.rs
@@ -1,10 +1,29 @@
 // SPDX-FileCopyrightText: Copyright (c) 2023-2025 Yegor Bugayenko
 // SPDX-License-Identifier: MIT
 
-use crate::SetDrain;
+use super::Set;
 use core::iter::FusedIterator;
 
-impl<K> Iterator for SetDrain<'_, K> {
+#[repr(transparent)]
+pub struct Drain<'a, T> {
+    iter: crate::map::drain::Drain<'a, T, ()>,
+}
+
+impl<T, const N: usize> Set<T, N> {
+    /// Clears the set, returning all elements as an iterator. Keeps the allocated
+    /// memory for reuse.
+    ///
+    /// If the returned iterator is dropped before being fully consumed, it drops the
+    /// remaining elements. The returned iterator keeps a mutable borrow on the set
+    /// to optimize its implementation.
+    pub fn drain(&mut self) -> Drain<'_, T> {
+        Drain {
+            iter: self.map.drain(),
+        }
+    }
+}
+
+impl<K> Iterator for Drain<'_, K> {
     type Item = K;
 
     #[inline]
@@ -18,11 +37,11 @@ impl<K> Iterator for SetDrain<'_, K> {
     }
 }
 
-impl<K> ExactSizeIterator for SetDrain<'_, K> {
+impl<K> ExactSizeIterator for Drain<'_, K> {
     #[inline]
     fn len(&self) -> usize {
         self.iter.len()
     }
 }
 
-impl<K> FusedIterator for SetDrain<'_, K> {}
+impl<K> FusedIterator for Drain<'_, K> {}
